muhinvgbcgxf

+++ PROGRAMACION/CAP12/mysqlCRUD-main/src/App.java
@@ -12,7 +12,7 @@ public class App {
         System.out.println("MENÚ PRINCIPAL");
         System.out.println("==============");
         System.out.println("1. Menú de grupos");
-        System.out.println("e. Menú de alumnos");
+        System.out.println("2. Menú de alumnos");
         System.out.println("3. Salir");
         while (true) {
             try {
@@ -83,6 +83,10 @@ public class App {
         String nombre = System.console().readLine();
         System.out.print("Introduzca el nombre del tutor: ");
         String profesor = System.console().readLine();
+        if (nombre.isEmpty() || profesor.isEmpty()) {
+            System.out.println("Los valores no pueden ser nulos");
+            return;
+        }
         try {
             Long id = gservice.create(new Grupo(0, nombre, profesor));    
             System.out.printf("Grupo creado correctamente (id: %d)\n", id);
@@ -101,10 +105,28 @@ public class App {
         }
         System.out.print("Elija el grupo a editar: ");
         Long id = Long.parseLong(System.console().readLine());
+        if (!comprobarExistenciaGrupoPorId(id, gservice)) {
+            System.out.println("Grupo no encontrado");
+            return;
+        }
         System.out.println("Introduzca el nombre del grupo: ");
         String nombre = System.console().readLine();
+        if (nombre.isEmpty()) {
+            try {
+                nombre = gservice.requestById(id).getNombre();
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+        }
         System.out.println("Introduzca el nombre del tutor: ");
         String profesor = System.console().readLine();
+        if (profesor.isEmpty()) {
+            try {
+                profesor = gservice.requestById(id).getProfesor();
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+        }
         try {
             int rowAffected = gservice.update(new Grupo(id, nombre, profesor));
             if(rowAffected == 1)
@@ -113,18 +135,22 @@ public class App {
                 System.out.println("No se ha podido actualizar el grupo");
         } 
         catch (SQLException e) {
-            System.out.println("No se ha podido actualizar el grupo");
             System.out.println("Ocurrión una excepción: "+ e.getMessage());
         }
     }
     private static void borrargrupo(GruposService gservice) {
         try {
             gservice.requestAll();
-        } catch (SQLException e) {
+        } 
+        catch (SQLException e) {
             e.printStackTrace();
         }
         System.out.print("Elija el grupo a borrar: ");
         Long id = Long.parseLong(System.console().readLine());
+        if (!comprobarExistenciaGrupoPorId(id, gservice)) {
+            System.out.println("Grupo no encontrado");
+            return;
+        }
         try {
             gservice.delete(id);
         } 
@@ -135,15 +161,16 @@ public class App {
     public static void listarGrupos(GruposService service){
         try {
             ArrayList<Grupo> grupos = service.requestAll();
-            if(grupos.size()==0){
+            if(grupos.size()==0) {
                 System.out.println("No hay grupos de alumnos");
             }
-            else{
+            else {
                 for(Grupo g : grupos){
                     System.out.println(g);
                 }
             }  
-        } catch (SQLException e) {
+        } 
+        catch (SQLException e) {
             e.printStackTrace();
         }
     }
@@ -155,27 +182,90 @@ public class App {
         }
         System.out.print("Elija el grupo a visualizar: ");
         Long id = Long.parseLong(System.console().readLine());
-        Grupo grupo;
+        if (!comprobarExistenciaGrupoPorId(id, gservice)) {
+            System.out.println("Grupo no encontrado");
+            return;
+        }
         try {
-            grupo = gservice.requestById(id);
-            if(grupo!=null)
-            System.out.println(grupo);
+            System.out.println(gservice.requestById(id));
         } catch (SQLException e) {
             e.printStackTrace();
         }
     }
-    private static void importarCSVGrupo() {
+    public static void importarCSVGrupo() {
 
     }
-    private static void exportarCSVGrupo() {
+    public static void exportarCSVGrupo() {
 
+    }
+    private static boolean comprobarExistenciaGrupoPorId(Long id, GruposService gservice) {
+        try {
+            if (gservice.requestById(id)==null) {
+                return false;
+            }
+        } 
+        catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return true;
     }
 
     //opciones menu alumno
-    public static void crearAlumno(AlumnosService aservice) {
-
+    public static void crearAlumno(AlumnosService aservice, GruposService gservice) {
+        System.out.print("\nIntroduzca el nombre del alumno: ");
+        String nombre = System.console().readLine();
+        System.out.print("Introduzca los apellidos del alumno: ");
+        String apellidos = System.console().readLine();
+        if (nombre.isEmpty() || apellidos.isEmpty()) {
+            System.out.println("Los valores no pueden ser nulos");
+            return;
+        }
+        System.out.print("Introduzca el código del grupo al que pertenece: ");
+        Long codigo = Long.parseLong(System.console().readLine());
+        if (!comprobarExistenciaGrupoPorId(codigo, gservice)) {
+            System.out.println("El código introducido no corresponde con ningun grupo existente");
+            return;
+        }
+        try {
+            Long codigoAlumno = aservice.create(new Alumno(0, nombre, apellidos, codigo));
+            System.out.printf("Alumno creado correctamente (id: %d)\n", codigoAlumno);
+        }
+        catch (SQLException e) {
+            e.printStackTrace();
+        }
     }
-    public static void editarAlumno(AlumnosService aservice) {
+    public static void editarAlumno(AlumnosService aservice, GruposService gservice) {
+        try {
+            aservice.requestAll();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        System.out.print("\nElija el alumno a editar: ");
+        Long id = Long.parseLong(System.console().readLine());
+        if (!comprobarExistenciaAlumnoPorId(id, aservice)) {
+            System.out.println("Alumno no encontrado");
+            return;
+        }
+        System.out.print("Introduzca el nombre del alumno: ");
+        String nombre = System.console().readLine();
+        if (nombre.isEmpty()) {
+            try {
+                nombre = aservice.requestById(id).getNombre();
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+        }
+        System.out.print("Introduzca los apellidos del alumno: ");
+        String apellidos = System.console().readLine();
+        if (apellidos.isEmpty()) {
+            try {
+                apellidos = aservice.requestById(id).getApellidos();
+            }
+            catch (SQLException e) {
+                e.printStackTrace();
+            }
+        }
+        System.out.print("Introduzca el código del grupo al que pretence: ");
 
     }
     public static void borrarAlumno(AlumnosService aservice) {
@@ -185,13 +275,44 @@ public class App {
 
     }
     public static void visualizarAlumno(AlumnosService aservice) {
+        try {
+            aservice.requestAll();
+        }
+        catch (SQLException e) {
+            e.printStackTrace();
+        }
+        System.out.print("\nElija el alumno a visualizar: ");
+        Long id = Long.parseLong(System.console().readLine());
+        try {
+            Alumno alumno = aservice.requestById(id);
+            if (alumno!=null) {
+                System.out.println(alumno);
+            }
+            else {
+                System.out.println("No se ha encontrado el alumno solicitado");
+            }
+        }
+        catch (SQLException e) {
+            e.printStackTrace();
+        }
+        
+    }
+    public static void importarCSVAlumno() {
 
     }
-    private static void importarCSVAlumno() {
+    public static void exportarCSVAlumno() {
 
     }
-    private static void exportarCSVAlumno() {
-
+    private static boolean comprobarExistenciaAlumnoPorId(Long id, AlumnosService aserive) {
+        try {
+            if (aserive.requestById(id)==null) {
+                return false;
+            }
+        }
+        catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return true;
     }
 
     public static void main(String[] args) throws Exception {
@@ -243,10 +364,10 @@ public class App {
                     while(!salirAlumno){
                         switch (menuAlumno()) {
                             case 1:
-                                crearAlumno(aservice);
+                                crearAlumno(aservice, gservice);
                                 break;
                             case 2:
-                                editarAlumno(aservice);
+                                editarAlumno(aservice, gservice);
                                 break;
                             case 3:
                                 borrarAlumno(aservice);
